Move CommunityMonitorModal into ModalQueue folder
