+++ front/app/containers/App/ModalQueue/modals/CommunityMonitorModal/CommunityMonitorModal.test.tsx
