+++ front/app/containers/App/ModalQueue/modals/CommunityMonitorModal/components/QuestionPreview.tsx
