+++ front/app/containers/App/ModalQueue/modals/CommunityMonitorModal/events.ts
