+++ front/app/containers/App/ModalQueue/modals/CommunityMonitorModal/index.tsx
